# mobile-edge/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SmartResponsor Mobile Edge
  version: v4
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health
      responses: { '200': { description: OK } }
  /config:
    get:
      summary: Config flags
      responses:
        '200':
          description: Flags
          headers:
            X-SR-Config-Signature: { description: HMAC hex, schema: { type: string } }
            X-SR-Secret-Version: { description: Secret version, schema: { type: string } }
          content:
            application/json:
              schema: { type: object }
  /entitlement:
    get:
      summary: Entitlement grants
      security: [ { bearerAuth: [] } ]
      responses: { '200': { description: Grants, content: { application/json: { schema: { type: object } } } } }
  /access/signin:
    post:
      summary: Proxy mobile access sign in to Accessing
      description: Accessing owns authentication and session transport; mobile-edge forwards request credentials and passes Accessing response cookies back to the mobile client without minting mobile tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileAccessSignInRequest'
      responses:
        '202':
          description: Accessing accepted the sign in and session transport is pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessSessionPayload'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
        '503':
          description: Accessing API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
  /access/register:
    post:
      summary: Proxy mobile access registration to Accessing
      description: Accessing owns registration-side session transport; mobile-edge passes Accessing response cookies back to the mobile client without minting mobile tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileAccessRegisterRequest'
      responses:
        '201':
          description: Accessing accepted the registration request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessSessionPayload'
        '409':
          description: Registration conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
        '503':
          description: Accessing API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
  /access/logout:
    post:
      summary: Proxy mobile access logout to Accessing
      description: Accessing owns logout and session clearing; mobile-edge passes Accessing response cookies back to the mobile client.
      responses:
        '200':
          description: Accessing returned the current session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessSessionPayload'
        '503':
          description: Accessing API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
  /access/session:
    get:
      summary: Proxy mobile access session to Accessing
      description: Accessing owns session lookup and transport; mobile-edge forwards client session credentials and passes Accessing response cookies back to the mobile client.
      responses:
        '200':
          description: Accessing returned the current mobile access session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessSessionPayload'
        '503':
          description: Accessing API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileAccessErrorPayload'
  /push/register:
    post:
      summary: Register push token
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushRegisterV1'
      responses: { '204': { description: No Content } }
  /receipt/verify:
    post:
      summary: Verify purchase receipt
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptVerifyV1'
      responses: { '200': { description: OK } }
  /analytic/batch:
    post:
      summary: Analytics JSON batch
      parameters:
        - in: header
          name: X-SR-Analytic-Signature
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnalyticBatchJsonV1' }
      responses: { '202': { description: Accepted } }
  /analytic/batch-ndjson:
    post:
      summary: Analytics NDJSON
      parameters:
        - in: header
          name: X-SR-Analytic-Signature
          required: true
          schema: { type: string }
      responses: { '202': { description: Accepted } }
  /sync/event:
    post:
      summary: Client sync event
      responses: { '204': { description: No Content } }
  /admin/stats:
    get:
      summary: Admin stats
      parameters:
        - in: header
          name: X-Admin-Token
          required: true
          schema: { type: string }
      responses: { '200': { description: OK } }
  /admin/schema/{name}/{ver}:
    get:
      summary: Schema fetch
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: path
          name: ver
          required: true
          schema: { type: string }
        - in: header
          name: X-Admin-Token
          required: true
          schema: { type: string }
      responses: { '200': { description: OK } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    PushRegisterV1:
      type: object
      required: [deviceId, token]
      properties:
        deviceId: { type: string }
        token: { type: string }
        platform: { type: string, enum: [android, ios, web] }
    MobileAccessSignInRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 8 }
    MobileAccessRegisterRequest:
      type: object
      required: [displayName, email, password]
      properties:
        displayName: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 8 }
    MobileAccessIdentityPayload:
      type: object
      additionalProperties: false
      description: accountId is a nullable compatibility field and is not sourced from Accessing in this milestone.
      required: [vendorId, accountId, displayName, email, emailVerified, secondFactorEnabled]
      properties:
        vendorId: { type: string }
        accountId:
          anyOf:
            - { type: string }
            - { type: 'null' }
        displayName: { type: string }
        email: { type: string }
        emailVerified: { type: boolean }
        secondFactorEnabled: { type: boolean }
    MobileAccessSessionPayload:
      type: object
      additionalProperties: false
      required: [status, identity, requiresVerification, requiresSecondFactor]
      properties:
        status: { type: string }
        identity:
          anyOf:
            - { $ref: '#/components/schemas/MobileAccessIdentityPayload' }
            - { type: 'null' }
        requiresVerification: { type: boolean }
        requiresSecondFactor: { type: boolean }
    MobileAccessErrorPayload:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fieldErrors:
          anyOf:
            - type: object
              additionalProperties:
                type: array
                items: { type: string }
            - { type: 'null' }
    ReceiptVerifyV1:
      type: object
      required: [transactionId, productId, platform, receipt]
      properties:
        transactionId: { type: string }
        productId: { type: string }
        platform: { type: string, enum: [android, ios] }
        receipt: { type: string }
    AnalyticBatchJsonV1:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            type: object
            required: [type, ts]
            properties:
              type: { type: string }
              ts: { type: number }